+++ src/homepage/src/screens/home/Animation/Background.js
@@ -14,7 +14,6 @@ const Container = styled.div`
   bottom: 0;
   right: 0;
   left: 0;
-  z-index: 0;
   pointer-events: none;
   ${delay};
 `;
