+++ src/homepage/src/screens/home/Animation/canvas/Dot.js
